+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning].
 - Add support for multiple cron expressions.
 - Inprove logs by always assigning the cron job name a non-null value.
 
+### Removed
+
+- Remove confusing overloads of the `AddCronJob` extension method.
+
 ## [1.1.0] - 2024-06-14
 
 ### Added