+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+## [2.0.0] - 2024-06-21
+
 ### Changed
 
 - Add support for multiple cron expressions.
@@ -41,6 +43,7 @@ and this project adheres to [Semantic Versioning].
 [Keep a Changelog]: https://keepachangelog.com/
 [Semantic Versioning]: https://semver.org/
 
-[Unreleased]: https://github.com/bfriesen/RandomSkunk.Hosting.Cron/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/bfriesen/RandomSkunk.Hosting.Cron/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/bfriesen/RandomSkunk.Hosting.Cron/compare/v1.1.0...v2.0.0
 [1.1.0]: https://github.com/bfriesen/RandomSkunk.Hosting.Cron/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/bfriesen/RandomSkunk.Hosting.Cron/compare/v0.0.0...v1.0.0